Move all game related code to game.js
